+++ src/Twig/Extension/I18n.php
@@ -3,9 +3,9 @@ declare(strict_types = 1);
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-namespace Wdes\PIL\Twig\Extension;
+namespace Wdes\phpI18nL10n\Twig\Extension;
 
-use Wdes\PIL\Twig\TokenParser;
+use Wdes\phpI18nL10n\Twig\TokenParser;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
@@ -25,7 +25,7 @@ class I18n extends AbstractExtension
     {
         return [
             new TwigFilter(
-                'trans', '\Wdes\PIL\Launcher::gettext'
+                'trans', '\Wdes\phpI18nL10n\Launcher::gettext'
             ),
         ];
     }