+++ src/plugins/MoReader.php
@@ -3,9 +3,9 @@ declare(strict_types = 1);
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-namespace Wdes\PIL\plugins;
+namespace Wdes\phpI18nL10n\plugins;
 
-use \Wdes\PIL\plugins\BasePlugin;
+use Wdes\phpI18nL10n\plugins\BasePlugin;
 use \Exception;
 use \stdClass;